+++ src/alb_sim/photon_mapping/build_photon_map_data.py
@@ -12,10 +12,18 @@ def build_photon_map_data(
 
     for photon_type in photon_storage.positions:
         maps[photon_type] = PhotonMapData(
-            positions=np.concatenate(photon_storage.positions[photon_type], axis=0),
-            directions=np.concatenate(photon_storage.directions[photon_type], axis=0),
-            energies=np.concatenate(photon_storage.energies[photon_type], axis=0),
-            times=np.concatenate(photon_storage.times[photon_type], axis=0),
+            positions=np.concatenate(
+                photon_storage.positions[photon_type], axis=0
+            ).astype(np.float32),
+            directions=np.concatenate(
+                photon_storage.directions[photon_type], axis=0
+            ).astype(np.float32),
+            energies=np.concatenate(
+                photon_storage.energies[photon_type], axis=0
+            ).astype(np.float32),
+            times=np.concatenate(photon_storage.times[photon_type], axis=0).astype(
+                np.float32
+            ),
         )
 
     return maps
